requirements.txt

Pillow==8.3.2
reportlab==3.5.68
